refactor: move extractFirstImage.ts
