+++ src/components/pages/posts/index/PostEditSection.tsx
@@ -10,7 +10,7 @@ import useNotification from "src/contexts/notification/useNotfication";
 import { clientAxios } from "src/utils/clientAxios";
 import useAuthSWR from "src/hooks/useAuthSWR";
 import ContentEditor from "src/components/common/ContentEditor";
-import extractFirstImage from "src/utils/extractFirstImage";
+import extractFirstImage from "src/utils/[slug]/extractFirstImage";
 import { AxiosError } from "axios";
 
 interface Props {
