+++ src/utils/[slug]/extractFirstImage.ts
